Rename feature aop to retry
In order to use the same feature through different implementation
